consolidate sftp util to adapter
