+++ cliboa/scenario/load/sftp.py
@@ -13,10 +13,10 @@
 #
 import os
 
+from cliboa.adapter.sftp import SftpAdapter
 from cliboa.core.validator import EssentialParameters
 from cliboa.scenario.sftp import BaseSftp
 from cliboa.util.constant import StepStatus
-from cliboa.util.sftp import Sftp
 
 
 class SftpBaseLoad(BaseSftp):
@@ -56,7 +56,7 @@ class SftpUpload(SftpBaseLoad):
                     self._logger.info("0 byte file will no be uploaded %s." % file)
                     continue
 
-                obj = Sftp().put_file(
+                obj = SftpAdapter().put_file(
                     src=file,
                     dest=os.path.join(self._dest_dir, os.path.basename(file)),
                     endfile_suffix=self._endfile_suffix,
